Feature/make some random changes

+++ server/src/seeds/cleanDb.ts
@@ -1,11 +1,22 @@
-import models from '../models/index.js';
-import db from '../config/connection.js';
+import models from "../models/index.js";
+import db from "../config/connection.js";
 
 export default async (modelName: "Question", collectionName: string) => {
   try {
-    let modelExists = await models[modelName].db.db.listCollections({
-      name: collectionName
-    }).toArray()
+    if (
+      !models ||
+      !models[modelName] ||
+      !models[modelName].db ||
+      !models[modelName].db.db
+    ) {
+      throw new Error(`Model ${modelName} does not exist`);
+    }
+
+    let modelExists = await models[modelName].db.db
+      .listCollections({
+        name: collectionName,
+      })
+      .toArray();
 
     if (modelExists.length) {
       await db.dropCollection(collectionName);
@@ -13,4 +24,4 @@ export default async (modelName: "Question", collectionName: string) => {
   } catch (err) {
     throw err;
   }
-}
+};
